20240525_code

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: rohta <rohta@student.42.jp>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 16:47:34 by rohta             #+#    #+#             */
-/*   Updated: 2024/05/14 18:10:19 by rohta            ###   ########.fr       */
+/*   Updated: 2024/05/25 16:04:17 by rohta            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (str[len] != '\0')
